#251 reorg indexed-db folder. Move the parent company functions to the company-icon folder
